Update camera_options_ZWO.json
Added concept of "headers" which help visually group options on web page.
Added a LOT of new options, and renamed some of the old ones to be more consistent.
Because there are more options, added the concept of "advanced" options which are hidden by default.  These options are not as likely to be used by the average person.
I took a first guess at what should be advanced, but this could use review.