Update liveview.php
Displays a message telling how often the image is updated for both daytime and nighttime images.
It references the old "delay" option (in addition to the new "daydelay" and "nightdelay" options).  This reference can be removed if the new camera options .json files are used (which use the new option names).