Update system.php
Allow displaying CPU temp in Celsius, Fahrenheit, or both.
Allow displaying fan speed if /home/pi/fan/fandata.txt exists.  The 1st (unused) number is the RPM; the 2nd number is the percent.  NOTE: please review the cutoffs for "danger", "warning", and "success".
De-activated the "progress-bar-striped" class since it eats CPU and if the user leave the page open for long, can increase the temperature.  NOTE: I prepended "ECC" to the class name which in effect de-activates it.  If you disagree with this change, simply remove "ECC".  If you agree with the change, you can remove "ECCprogress-bar-striped" in multiple locations.  The impact of not using that class is you don't get the moving striped bars (you still get bars though).