Eric Claeys' changes to GUI (#20)
* Update index.php

* Update public.php

* Update custom.css

Added CSS for headers and advanced options on GUI admin page.

* Update camera_options_ZWO.json

Added concept of "headers" which help visually group options on web page.
Added a LOT of new options, and renamed some of the old ones to be more consistent.
Because there are more options, added the concept of "advanced" options which are hidden by default.  These options are not as likely to be used by the average person.
I took a first guess at what should be advanced, but this could use review.

* Update camera_options_RPiHQ.json

Added concept of "headers" to group options to aid in visibility.
Added concept of "advanced" options.  NEED HELP HERE DETERMINING WHICH OPTIONS SHOULD BE ADVANCED, IF ANY.

* Update camera_options_ZWO.json

For consistency with the Pi camera options file, put the "advanced" entries at the end.

* Create dashboard_eth0.php

Adds ability to view/control the LAN interface, similar to WLAN page.
NOTE: since the start/stop eth0 button needs testing, there's a "useEth0" variable in index.php that's currently set to "false".  After testing, this should be set to "true", or the "if (useEth0)" statements removed.

* Update camera_settings.php

Implements headers to visually group options.
Implements advanced options and the ability to show/hide them.  They are highlighted using CSS in dist/css/custom.css.
Allows single quotes in values and descriptions.
Use tables to display each option for better control over spacing.

* Update days.php

Allow showing All timelapse.
Changed "all" to "All" in URLs so the startrails, keogram, and timelapse page headers would look nicer.

* Update functions.php

Added get_variable() to get a variable from a file, since its functionality was used in multiple other files.
Added ListFileType() to reduce 99% redundant code in keograms.php and startrails.php.  Also merged in code from videos.php so all three files are now simply calls to ListFileType().  Those files could, in theory, be deleted and their one line moved into index.php.

* Update images.php

Informs user when there are no images, rather than leaving a mostly blank screen.

* Update keograms.php

The functionality that was here is now in ListFileType() in functions.php

* Update startrails.php

The functionality that was in this file is now in ListFileType() in functions.php.

* Update liveview.php

Displays a message telling how often the image is updated for both daytime and nighttime images.
It references the old "delay" option (in addition to the new "daydelay" and "nightdelay" options).  This reference can be removed if the new camera options .json files are used (which use the new option names).

* Update system.php

Allow displaying CPU temp in Celsius, Fahrenheit, or both.
Allow displaying fan speed if /home/pi/fan/fandata.txt exists.  The 1st (unused) number is the RPM; the 2nd number is the percent.  NOTE: please review the cutoffs for "danger", "warning", and "success".
De-activated the "progress-bar-striped" class since it eats CPU and if the user leave the page open for long, can increase the temperature.  NOTE: I prepended "ECC" to the class name which in effect de-activates it.  If you disagree with this change, simply remove "ECC".  If you agree with the change, you can remove "ECCprogress-bar-striped" in multiple locations.  The impact of not using that class is you don't get the moving striped bars (you still get bars though).

* Update videos.php

The functionality that was in this file is now in ListFileType() in functions.php.
Additionally, timelapse for ALL days can now be shown.

* Update index.php

Uses get_variable() in functions.php.
Optionally includes the new dashboard_eth0.php file.
Uses the new IMG_DIR and IMG_PREFIX variables to allow users to move the image file and change the default "liveview-" prefix.  Calls to DisplayLiveView() now pass in the image name.
Added "<style>" section to line things up better.

* Update public.php

Uses get_variable() in functions.php.
Uses new IMG_DIR and IMG_PREFIX variables from config.sh to allow user to specify different directory and prefix for image.

* Update camera_options_RPiHQ.json

Updated variable names to be consistent with the ZWO version.

* Update camera_options_RPiHQ.json

* Update camera_options_ZWO.json

Renamed "time" to "showTime" to be consistent with other "show" variables.

* Added more Pi models

Added extra Pi models from astroimagineer.