safer date filter
instead of filtering out a few potentially bad directory names, only
include acceptable names. A regex match for 8 digits where the first
one is '2' comes pretty close. Sure, it would permit the 42nd day of
the 13th month of 2999, but it's better than listing random stuff in
the images directory.