dashboard.php: improve Link Quality
The Link Quality returned from iwconfig contains two numbers, the absolute value and the max value.  We were previously using the absolute value as the percent, which was incorrect.  The percent is  (absolute value / max value).  This commit fixes that, and uses colors to indicate issues as is done on the "System" page.
Note that the color cutoffs are my best guess.